+++ cli/src/enums/mod.rs
@@ -1,9 +1,9 @@
-pub mod calendar_scope;
+pub mod calendar_context;
 pub mod easter_calculation_type;
 pub mod output_format;
 pub mod validation_type;
 
-pub use calendar_scope::CliCalendarScope;
+pub use calendar_context::CliCalendarContext;
 pub use easter_calculation_type::CliEasterCalculationType;
 pub use output_format::{CliOutputFormat, OutputFormat};
 pub use validation_type::ValidationType;