+++ core/src/types/config/mod.rs
@@ -1,7 +1,7 @@
-pub mod calendar_scope;
+pub mod calendar_context;
 pub mod easter_calculation_type;
 pub mod particular_config;
 
-pub use calendar_scope::*;
+pub use calendar_context::*;
 pub use easter_calculation_type::*;
 pub use particular_config::*;